Initial separation, works on local deployments, needs cleanup
Finally isolated and fixed the test failure
