+++ SpatialGDK/Source/SpatialGDK/Public/Interop/Connection/SpatialWorkerConnection.h
@@ -7,12 +7,9 @@
 #include "HAL/ThreadSafeBool.h"
 
 #include "Interop/Connection/SpatialOSWorkerInterface.h"
-#include "Interop/Connection/ConnectionConfig.h"
 #include "Interop/Connection/OutgoingMessages.h"
 #include "SpatialCommonTypes.h"
-#include "SpatialGDKSettings.h"
 #include "UObject/WeakObjectPtr.h"
-#include "Utils/SpatialLatencyTracer.h"
 
 #include <WorkerSDK/improbable/c_schema.h>
 #include <WorkerSDK/improbable/c_worker.h>
@@ -21,33 +18,15 @@
 
 DECLARE_LOG_CATEGORY_EXTERN(LogSpatialWorkerConnection, Log, All);
 
-enum class ESpatialConnectionType
-{
-	Receptionist,
-	LegacyLocator,
-	Locator,
-	DevAuthFlow
-};
-
 UCLASS()
 class SPATIALGDK_API USpatialWorkerConnection : public UObject, public FRunnable, public SpatialOSWorkerInterface
 {
 	GENERATED_BODY()
 
 public:
+	void SetConection(Worker_Connection* WorkerConnectionIn);
 	virtual void FinishDestroy() override;
 	void DestroyConnection();
-	
-	using LoginTokenResponseCallback = TFunction<bool(const Worker_Alpha_LoginTokensResponse*)>;
-    
-    /// Register a callback using this function.
-    /// It will be triggered when receiving login tokens using the development authentication flow inside SpatialWorkerConnection.
-    /// @param Callback - callback function.
-	void RegisterOnLoginTokensCallback(const LoginTokenResponseCallback& Callback) {LoginTokenResCallback = Callback;}
-
-	void Connect(bool bConnectAsClient, uint32 PlayInEditorID);
-
-	FORCEINLINE bool IsConnected() { return bIsConnected; }
 
 	// Worker Connection Interface
 	virtual TArray<Worker_OpList*> GetOpList() override;
@@ -68,42 +47,16 @@ public:
 	PhysicalWorkerName GetWorkerId() const;
 	const TArray<FString>& GetWorkerAttributes() const;
 
-	void SetConnectionType(ESpatialConnectionType InConnectionType);
-
-	// TODO: UNR-2753
-	FReceptionistConfig ReceptionistConfig;
-	FLocatorConfig LocatorConfig;
-	FDevAuthConfig DevAuthConfig;
-
 	DECLARE_MULTICAST_DELEGATE_OneParam(FOnEnqueueMessage, const SpatialGDK::FOutgoingMessage*);
 	FOnEnqueueMessage OnEnqueueMessage;
 
 	DECLARE_MULTICAST_DELEGATE_OneParam(FOnDequeueMessage, const SpatialGDK::FOutgoingMessage*);
 	FOnDequeueMessage OnDequeueMessage;
 
-	DECLARE_DELEGATE(OnConnectionToSpatialOSSucceededDelegate)
-	OnConnectionToSpatialOSSucceededDelegate OnConnectedCallback;
-
-	DECLARE_DELEGATE_TwoParams(OnConnectionToSpatialOSFailedDelegate, uint8_t, const FString&);
-	OnConnectionToSpatialOSFailedDelegate OnFailedToConnectCallback;
-
-	bool TrySetupConnectionConfigFromCommandLine(const FString& SpatialWorkerType);
-	void SetupConnectionConfigFromURL(const FURL& URL, const FString& SpatialWorkerType);
-	void RequestDeploymentLoginTokens();
-
 	void QueueLatestOpList();
 	void ProcessOutgoingMessages();
 
 private:
-	void ConnectToReceptionist(uint32 PlayInEditorID);
-	void ConnectToLocator(FLocatorConfig* InLocatorConfig);
-	void FinishConnecting(Worker_ConnectionFuture* ConnectionFuture);
-
-	void OnConnectionSuccess();
-	void OnConnectionFailure();
-
-	ESpatialConnectionType GetConnectionType() const;
-
 	void CacheWorkerAttributes();
 
 	// Begin FRunnable Interface
@@ -114,20 +67,11 @@ private:
 
 	void InitializeOpsProcessingThread();
 
-	void StartDevelopmentAuth(const FString& DevAuthToken);
-	static void OnPlayerIdentityToken(void* UserData, const Worker_Alpha_PlayerIdentityTokenResponse* PIToken);
-	static void OnLoginTokens(void* UserData, const Worker_Alpha_LoginTokensResponse* LoginTokens);
-	void ProcessLoginTokensResponse(const Worker_Alpha_LoginTokensResponse* LoginTokens);
-
 	template <typename T, typename... ArgsType>
 	void QueueOutgoingMessage(ArgsType&&... Args);
 
 private:
 	Worker_Connection* WorkerConnection;
-	Worker_Locator* WorkerLocator;
-
-	bool bIsConnected;
-	bool bConnectAsClient = false;
 
 	TArray<FString> CachedWorkerAttributes;
 
@@ -140,7 +84,4 @@ private:
 
 	// RequestIds per worker connection start at 0 and incrementally go up each command sent.
 	Worker_RequestId NextRequestId = 0;
-
-	ESpatialConnectionType ConnectionType = ESpatialConnectionType::Receptionist;
-	LoginTokenResponseCallback LoginTokenResCallback;
 };
